bigint 수정

+++ src/migrations/20221216121313-create-posts.js
@@ -42,7 +42,7 @@ module.exports = {
                 allowNull: false,
             },
             recruitmentEndDay: {
-                type: Sequelize.DataTypes.INTEGER,
+                type: Sequelize.DataTypes.BIGINT,
                 allowNull: false,
             },
             startTime: {
